// o/src/app/api/app.generated.ts
export class EventDto implements IEventDto {
    id!: number;
    title?: string | undefined;
    date!: Date;
    description?: string | undefined;

    constructor(data?: IEventDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["id"];
            this.title = data["title"];
            this.date = data["date"] ? new Date(data["date"].toString()) : <any>undefined;
            this.description = data["description"];
        }
    }

    static fromJS(data: any): EventDto {
        data = typeof data === 'object' ? data : {};
        let result = new EventDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["title"] = this.title;
        data["date"] = this.date ? this.date.toISOString() : <any>undefined;
        data["description"] = this.description;
        return data; 
    }
}

export interface IEventDto {
    id: number;
    title?: string | undefined;
    date: Date;
    description?: string | undefined;
}